Die on connection unavailable
Sometimes connection closing is masked as connection unavailable, when this occurs the agent never recovers.
Taking `codes.Unavailable` out of the allowed errors so that when the agent hits this it will die.

As the agent is ran as a docker container, docker can be used to restart it and recover.